requirements.txt
easyocr~=1.7.1
keyboard~=0.13.5
opencv_contrib_python~=4.8.0.76
opencv_python~=4.8.0.76
opencv_python_headless~=4.8.0.76
Pillow~=10.0.0
pywin32~=306
